Implement password salting for increased security
A salt was added to the user password hashing process in the sign-up and login feature to increase the account security; preventing matches when the same user passwords are hashed. Salt is being stored in the user database. Salt is combined with user password before being hashed. This salted password hash is stored and then compared with the newly hashed version during login.